+++ source/codec/decompressor.h
@@ -46,11 +46,11 @@ public:
     // needed.
     //
     virtual bool Process(const uint8_t* input_data,
-                         int input_size,
+                         size_t input_size,
                          uint8_t* output_data,
-                         int output_size,
-                         int* consumed,
-                         int* written) = 0;
+                         size_t output_size,
+                         size_t* consumed,
+                         size_t* written) = 0;
 };
 
 } // namespace aspia
