+++ source/codec/decompressor_zlib.h
@@ -25,11 +25,11 @@ public:
 
     // Decompressor implementations.
     bool Process(const uint8_t* input_data,
-                 int input_size,
+                 size_t input_size,
                  uint8_t* output_data,
-                 int output_size,
-                 int* consumed,
-                 int* written) override;
+                 size_t output_size,
+                 size_t* consumed,
+                 size_t* written) override;
 
 private:
     z_stream stream_;
